+++ src/vsmtp/vsmtp-config/src/tests/root_example/logging.rs
@@ -19,11 +19,13 @@ use vsmtp_common::{collection, CodeID, Reply, ReplyCode};
 
 #[test]
 fn parse() {
-    let toml = include_str!("../../../../../../examples/config/logging.toml");
     pretty_assertions::assert_eq!(
-        Config::from_toml(toml).unwrap(),
+        Config::from_vsl_file(std::path::PathBuf::from_iter([
+            env!("CARGO_MANIFEST_DIR"),
+            "../../../examples/config/logging.vsl"
+        ])).unwrap(),
         Config::builder()
-            .with_version_str(">=1.3.0-rc.0, <2.0.0")
+            .with_version_str(">=1.3.3, <2.0.0")
             .unwrap()
             .with_hostname()
             .with_default_system()