+++ src/vsmtp/vsmtp-core/src/main.rs
@@ -59,12 +59,7 @@ fn try_main() -> anyhow::Result<()> {
 
     let config = args.config.as_ref().map_or_else(
         || Ok(Config::default()),
-        |config| {
-            std::fs::read_to_string(config)
-                .context(format!("Cannot read file '{}'", config))
-                .and_then(|f| Config::from_toml(&f).context("File contains format error"))
-                .context("Cannot parse the configuration")
-        },
+        |path| Config::from_vsl_file(&path).context("Cannot parse the configuration"),
     )?;
 
     if let Some(command) = args.command {